Removed Nose dependency from Master and fixed imports.
Moved docstring to PythonNose module as it is the only location to
use the methods within. Fixed imports with DatabaseReader. Remove
xml hash files as they are not needed.

Proved that Nose finds unwanted tests and replaced Nose with the
methods found within docstrings. When discovering tests the import
source code is loaded within a child process to stop conflicts with
the YATES's code and the pollution of sys.path. Each module that
might conflict with the loaded source code is reloaded after the
sys.path includes the new source location.